feat: 전체 스크롤 영역 구현 (#33)
